+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 5.5.0
+
+* Add the value argument for `PickerLinkageEntry`, so `MultiColumnLinkagePicker` gets the `onValueChanged` method
+* `AreaPicker` removed. Use `MultiColumnLinkagePicker`, as shown in the `MultiColumnLinkagePicker` component in the example,
+* In `example/lib/module/picker/area.dart` The regional data in the dart
+
 ## 5.3.2
 
 * Fixed some issues with `DateTimePicker` and `AreaPicker`